api/requirements.txt
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Flask==3.0.3
Flask-JWT-Extended==4.6.0
google-api-core==2.21.0
google-auth==2.35.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
googleapis-common-protos==1.65.0
grpcio==1.67.0
grpcio-status==1.67.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.1
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyJWT==2.9.0
requests==2.32.3
rsa==4.9
urllib3==2.2.3
Werkzeug==3.0.4
